Slight changes to profile button location
